refactor out common code in embed role & directive
